refactor(subsonic): extract response builders into routes/subsonic/builders.ts (refs #243)
Phase 1 of 4: move buildSong/buildAlbum/annotateStarred/annotatePlays,
contentTypeForFormat, audioSourceFields, and the share-ID pickers plus
their row types (TrackRow, ReleaseGroupRow) out of the subsonic route
monolith. Pure move — no logic, signature, or behavior changes.
